add macro for test-template and remove duplicate code in test-curves
